+++ c-expr-dsl/test/Main.hs
@@ -2,8 +2,8 @@ module Main (main) where
 
 import Test.Tasty
 
-import Test.CExpr.Parse       qualified as Parse
-import Test.CExpr.Typecheck   qualified as Typecheck
+import Test.CExpr.Parse qualified as Parse
+import Test.CExpr.Typecheck qualified as Typecheck
 
 main :: IO ()
 main = defaultMain $ testGroup "c-expr-dsl" [