+++ hs-bindgen/src-internal/HsBindgen/Frontend/AST/Deps.hs
@@ -5,14 +5,14 @@ module HsBindgen.Frontend.AST.Deps (
   , depsOfUnion
   , depsOfField
     -- * Parsed macros
-  , MacroNameResolver
   , depsOfDeclParsedMacro
   ) where
 
 import Data.Set qualified as Set
 import GHC.Records
 
-import C.Expr.Syntax qualified as CExpr.DSL
+import C.Expr.Syntax qualified as CExpr
+import C.Expr.Typecheck qualified as CExpr
 import C.Expr.Typecheck.Interface.Value qualified as V
 
 import HsBindgen.Frontend.AST.Decl qualified as C
@@ -79,25 +79,28 @@ depsOfDeclTcMacro = depsOfDeclWith (depsOfTcMacro (Proxy @p))
 
 -- | Dependencies of typechecked macro declarations
 depsOfTcMacro ::
-     (IsPass p, MacroBody p ~ CheckedMacro p)
+     forall p. (IsPass p, MacroBody p ~ CheckedMacro p)
   => Proxy p -> MacroBody p -> [(ValOrRef, Id p)]
 depsOfTcMacro proxy = \case
-    MacroType typ  -> depsOfType typ.typ
-    MacroExpr expr -> depsOfVExpr proxy (Set.fromList expr.args) expr.body
+    -- TODO <https://github.com/well-typed/hs-bindgen/issues/1953>
+    --
+    -- If we replace the CType with the TExpr, we need to know if the
+    -- dependencies are by value or by reference.
+    MacroType  typ -> depsOfType  typ.cType
+    MacroValue val -> depsOfValue val.value
   where
+    depsOfValue :: CExpr.CheckedMacroValueExpr (Id p) -> [(ValOrRef, Id p)]
+    depsOfValue (CExpr.CheckedMacroValueExpr _ body _) = depsOfVExpr proxy body
+
     -- Collect value-level dependencies from a checked macro body. Local
     -- arguments (lambda-bound ids) are excluded.
     depsOfVExpr ::
-         forall p. (Ord (Id p))
-      => Proxy p
-      -> Set (Id p)
-      -> V.Expr (Id p)
+         forall ctx.
+         Proxy p
+      -> V.Expr ctx (Id p)
       -> [(ValOrRef, Id p)]
-    depsOfVExpr _ localArgs =
-        map (ByValue,) . filter (not . isLocalArg) . toList
-      where
-        isLocalArg :: Id p -> Bool
-        isLocalArg x = Set.member x localArgs
+    depsOfVExpr _ =
+        map (ByValue,) . toList
 
 {-------------------------------------------------------------------------------
   Structs and unions
@@ -133,65 +136,88 @@ depsOfTypedef typedef = depsOfType typedef.typ
 --   bare names (see 'MacroNameResolver').
 depsOfDeclParsedMacro ::
      forall p. (IsPass p, MacroBody p ~ ParsedMacro, Id p ~ DeclId)
-  => MacroNameResolver -> C.DeclKind p -> [(ValOrRef, Id p)]
-depsOfDeclParsedMacro resolver = depsOfDeclWith depsOfParsedMacro
+  => Set DeclId -> C.DeclKind p -> [(ValOrRef, Id p)]
+depsOfDeclParsedMacro allDeclIds = depsOfDeclWith depsOfParsedMacro
   where
     depsOfParsedMacro :: ParsedMacro -> [(ValOrRef, DeclId)]
-    depsOfParsedMacro (ParsedMacro m) = depsOfCExprMacro resolver m
-
--- | Resolves a bare name found in a parsed macro body
---
--- Result:
---
--- - @'Just' 'CNameKind'@: This bare name refers to another declaration that is in
---   the 'DeclIndex'. For example, a macro, or a @typedef@.
---
--- - @'Nothing'@: This bare name refers to a built-in type or a system @typedef@
---   not present in the 'DeclIndex'. In this case, we /do not generate/ a
---   dependency.
-type MacroNameResolver = Text -> Maybe CNameKind
+    depsOfParsedMacro (ParsedMacro m) = depsOfCExprMacro allDeclIds m
 
 -- | Collect all references in a macro body.
 --
 -- Local macro arguments are excluded.
 depsOfCExprMacro ::
-     MacroNameResolver
-  -> CExpr.DSL.Macro
+     Set DeclId
+  -> CExpr.Macro
   -> [(ValOrRef, DeclId)]
-depsOfCExprMacro resolver macro =
-    map (ByValue,) $ goExpr (Set.fromList macro.macroArgs) macro.macroExpr
+depsOfCExprMacro allDeclIds (CExpr.Macro _ _ _ macroExpr) =
+    map (ByValue,) $ goExpr macroExpr
   where
-    goExpr :: Set CExpr.DSL.Name -> CExpr.DSL.Expr CExpr.DSL.Ps -> [DeclId]
-    goExpr localArgs = \case
-      CExpr.DSL.Term  term   -> goTerm  localArgs term
-      CExpr.DSL.TyApp _ xs   -> concatMap (goExpr localArgs) xs
-      CExpr.DSL.VaApp _ _ xs -> concatMap (goExpr localArgs) xs
+    goExpr :: CExpr.Expr ctx CExpr.Ps -> [DeclId]
+    goExpr = \case
+      CExpr.Term  term   -> goTerm term
+      CExpr.TyApp _ xs   -> concatMap goExpr xs
+      CExpr.VaApp _ _ xs -> concatMap goExpr xs
 
-    goTerm :: Set CExpr.DSL.Name -> CExpr.DSL.Term CExpr.DSL.Ps -> [DeclId]
-    goTerm localArgs = \case
-      CExpr.DSL.Literal lit -> goLit lit
+    goTerm :: CExpr.Term ctx CExpr.Ps -> [DeclId]
+    goTerm = \case
+      CExpr.Literal lit ->
+        goLit lit
+      CExpr.LocalParam{} ->
+        []
       -- Variable / function call. A bare identifier is always parsed as Var;
       -- the typechecker decides whether it is a type or value reference.
       -- We use the resolver to emit a dep of the correct kind, or skip the
       -- name entirely if it is not a known declaration (e.g. a built-in type).
-      CExpr.DSL.Var _ nm callArgs
-        | nm `Set.member` localArgs ->
-            concatMap (goExpr localArgs) callArgs
-        | Just kind <- resolver nm.getName ->
-            mkId nm.getName kind :
-            concatMap (goExpr localArgs) callArgs
+      CExpr.Var _ nm callArgs
+        | Just declId <- resolveMacroTypedef nm.getName ->
+            declId : concatMap goExpr callArgs
         | otherwise ->
-            concatMap (goExpr localArgs) callArgs
+            concatMap goExpr callArgs
 
-    goLit :: CExpr.DSL.Literal -> [DeclId]
+    goLit :: CExpr.Literal -> [DeclId]
     goLit = \case
       -- Named type specifier using an elaborated tag: struct/union/enum.
-      CExpr.DSL.TypeLit (CExpr.DSL.TypeTagged tag nm) ->
-        [ mkId nm (CNameKindTagged (convertTagKind tag)) ]
+      CExpr.TypeTagged tag nm
+        | Just declId <- resolveTaggedType (convertTagKind tag) nm.getName ->
+            [declId]
+        | otherwise ->
+            []
       -- Other built-in type specifiers (int, char, etc.) have no dependencies.
-      CExpr.DSL.TypeLit _ -> []
+      CExpr.TypeLit{}  -> []
       -- Value literals have no dependencies.
-      _ -> []
+      CExpr.ValueLit{} -> []
 
-    mkId :: Text -> CNameKind -> DeclId
-    mkId name kind = DeclId (CDeclName name kind) False
+    -- TODO <https://github.com/well-typed/hs-bindgen/issues/1952>
+    --
+    -- We should only resolve the declaration IDs of macro dependencies one. Now
+    -- we resolve them twice: when we get the dependencies of parsed macros, and
+    -- when we typecheck macros.
+
+    -- | Resolves a bare name found in a parsed macro body
+    --
+    -- Result:
+    --
+    -- - @'Just' 'CNameKind'@: This bare name refers to another declaration that is in
+    --   the 'DeclIndex'. For example, a macro, or a @typedef@.
+    --
+    -- - @'Nothing'@: This bare name is not in the set of known declarations. It
+    --   may simply be non-existent, or refer to a built-in type or a system
+    --   @typedef@ not present in the 'DeclIndex'. In this case, we /do not
+    --   generate/ a dependency.
+    resolveMacroTypedef :: Text -> Maybe DeclId
+    resolveMacroTypedef nm
+      | macroId   `Set.member` allDeclIds = Just macroId
+      | typedefId `Set.member` allDeclIds = Just typedefId
+      | otherwise                         = Nothing
+      where
+        macroId, typedefId :: DeclId
+        macroId   = DeclId (CDeclName nm CNameKindMacro)    False
+        typedefId = DeclId (CDeclName nm CNameKindOrdinary) False
+
+    resolveTaggedType :: CTagKind -> Text -> Maybe DeclId
+    resolveTaggedType tag nm
+      | taggedId `Set.member` allDeclIds = Just taggedId
+      | otherwise                        = Nothing
+      where
+        taggedId :: DeclId
+        taggedId = DeclId (CDeclName nm $ CNameKindTagged tag) False