+++ hs-bindgen/src-internal/HsBindgen/TraceMsg.hs
@@ -14,8 +14,8 @@ module HsBindgen.TraceMsg (
   , ImmediateAssignAnonIdsMsg (..)
   , ImmediateParseMsg(..)
   , DelayedParseMsg(..)
-  , CExpr.DSL.MacroParseError(..)
-  , CExpr.DSL.MacroTcError(..)
+  , CExpr.MacroParseError(..)
+  , CExpr.MacroTcError(..)
   , ResolveBindingSpecsMsg(..)
   , ResolveHeaderMsg(..)
   , SelectMsg(..)
@@ -26,8 +26,8 @@ module HsBindgen.TraceMsg (
 
 import Data.List qualified as List
 
-import C.Expr.Parse qualified as CExpr.DSL
-import C.Expr.Typecheck qualified as CExpr.DSL
+import C.Expr.Parse qualified as CExpr
+import C.Expr.Typecheck qualified as CExpr
 
 import Clang.HighLevel.Types (Diagnostic (..))